+++ concordium-consensus/tests/globalstate/GlobalStateTests/ConfigureValidator.hs
@@ -105,10 +105,12 @@ testAddValidatorAllCases ::
     Spec
 testAddValidatorAllCases spv = describe "bsoAddValidator" $ do
     forM_ validatorConditions $ \vc -> do
-        it (show vc) $ runTest False vc
-        when supportCooldown $ it (show vc <> " with cooldown") $ runTest True vc
+        it (show vc) $ runTest False False vc
+        when supportCooldown $ it (show vc <> " with cooldown") $ runTest True False vc
+        when supportSuspension $ it (show vc <> " with suspended validator") $ runTest True True vc
   where
     supportCooldown = supportsFlexibleCooldown $ accountVersionFor $ demoteProtocolVersion (protocolVersion @pv)
+    supportSuspension = supportsValidatorSuspension $ accountVersionFor $ demoteProtocolVersion (protocolVersion @pv)
     minEquity = 1_000_000_000
     chainParams =
         DummyData.dummyChainParameters @(ChainParametersVersionFor pv)
@@ -129,7 +131,7 @@ testAddValidatorAllCases spv = describe "bsoAddValidator" $ do
                 DummyData.dummyArs
                 (withIsAuthorizationsVersionForPV spv DummyData.dummyKeyCollection)
                 chainParams
-    runTest withCooldown ValidatorConditions{..} = runTestBlockState @pv $ do
+    runTest withCooldown suspended ValidatorConditions{..} = runTestBlockState @pv $ do
         let va =
                 ValidatorAdd
                     { vaKeys = if vcAggregationKeyDuplicate then badKeys else goodKeys,
@@ -142,7 +144,8 @@ testAddValidatorAllCases spv = describe "bsoAddValidator" $ do
                             { _finalizationCommission = makeAmountFraction $ if vcFinalizationRewardNotInRange then 100 else 300,
                               _bakingCommission = makeAmountFraction $ if vcBakingRewardNotInRange then 100 else 500,
                               _transactionCommission = makeAmountFraction $ if vcTransactionFeeNotInRange then 300 else 100
-                            }
+                            },
+                      vaSuspended = suspended
                     }
         initialAccounts <- mapM makeDummyAccount (addValidatorTestAccounts withCooldown)
         initialBS <- mkInitialState initialAccounts
@@ -189,7 +192,7 @@ testAddValidatorAllCases spv = describe "bsoAddValidator" $ do
                                           _poolMetadataUrl = vaMetadataURL va,
                                           _poolCommissionRates = vaCommissionRates va
                                         },
-                                  _bieIsSuspended = conditionally (sSupportsValidatorSuspension (accountVersion @(AccountVersionFor pv))) False
+                                  _bieIsSuspended = conditionally (sSupportsValidatorSuspension (accountVersion @(AccountVersionFor pv))) suspended
                                 }
                         }
             bkr <- getAccountBaker (fromJust acc)
@@ -488,6 +491,11 @@ runUpdateValidatorTest spv commissionRanges ValidatorUpdateConfig{vucValidatorUp
                 forM_ (vuTransactionFeeCommission vu) $ \fee -> tell [BakerConfigureTransactionFeeCommission fee]
                 forM_ (vuBakingRewardCommission vu) $ \fee -> tell [BakerConfigureBakingRewardCommission fee]
                 forM_ (vuFinalizationRewardCommission vu) $ \fee -> tell [BakerConfigureFinalizationRewardCommission fee]
+                forM_ (vuSuspend vu) $ \suspended ->
+                    tell
+                        [ if suspended then BakerConfigureSuspended else BakerConfigureResumed
+                          | STrue <- [hasValidatorSuspension]
+                        ]
                 forM_ (vuCapital vu) $ \capital ->
                     tell $
                         if capital >= initialStakedAmount
@@ -526,6 +534,9 @@ runUpdateValidatorTest spv commissionRanges ValidatorUpdateConfig{vucValidatorUp
                         .~ (if newCapital == 0 then RemoveStake else ReduceStake newCapital)
                             (PendingChangeEffectiveV1 (24 * 60 * 60 * 1000 + 1000))
                 | otherwise = stakedAmount .~ newCapital
+        let updateSuspended suspend
+                | STrue <- hasValidatorSuspension = accountBakerInfo . bieIsSuspended .~ suspend
+                | otherwise = id
         let expectedAccountBaker
                 | STrue <- flexibleCooldown, vuCapital vu == Just 0 = Nothing
                 | otherwise =
@@ -546,10 +557,12 @@ runUpdateValidatorTest spv commissionRanges ValidatorUpdateConfig{vucValidatorUp
                             & maybe id (poolCommissionRates . finalizationCommission .~) (vuFinalizationRewardCommission vu)
                             & maybe id (poolCommissionRates . bakingCommission .~) (vuBakingRewardCommission vu)
                             & maybe id (poolCommissionRates . transactionCommission .~) (vuTransactionFeeCommission vu)
+                            & maybe id updateSuspended (vuSuspend vu)
         actualAccountBaker <- getAccountBaker acc0
         liftIO $ actualAccountBaker `shouldBe` expectedAccountBaker
   where
     flexibleCooldown = sSupportsFlexibleCooldown (accountVersion @(AccountVersionFor pv))
+    hasValidatorSuspension = sSupportsValidatorSuspension (accountVersion @(AccountVersionFor pv))
     minEquity = 1_000_000_000
     chainParams =
         DummyData.dummyChainParameters @(ChainParametersVersionFor pv)
@@ -577,3 +590,7 @@ tests lvl = parallel $ describe "Validator" $ do
         testAddValidatorAllCases SP7
         testUpdateValidator SP7 (lvl > 1)
         testUpdateValidatorOverlappingCommissions SP7
+    describe "P8" $ do
+        testAddValidatorAllCases SP8
+        testUpdateValidator SP8 (lvl > 1)
+        testUpdateValidatorOverlappingCommissions SP8